+++ filterseqscommand.h
@@ -24,9 +24,9 @@ public:
 	
 private:
 	struct linePair {
-		int start;
+		unsigned long int start;
 		int num;
-		linePair(long int i, long int j) : start(i), num(j) {}
+		linePair(unsigned long int i, long int j) : start(i), num(j) {}
 	};
 	vector<linePair*> lines;
 	vector<int> processIDS;
